tradesparkChallangeDjango/requirements.txt
Django==2.0.2
djangorestframework==3.9.3
django-cors-headers==2.1.0